remove the main menu from the UI and related code
The main menu was exposing functionality accessible through the
toolbar UI. Since the UI is moving toward a more modern design
(c.f. the GNOME3 Music application/view), the menu bar has been
removed.

The only bit that is not exposed is the view switching. As a
result, the notebook tabs have been enabled temporarily until the
correct navigation is in place.