add a 10px menu bar at the top
This is a workaround for when the titlebar is hidden in Gtk+ 3.3+.
When the titlebar is hidden, it is not possible to grab the window
with the mouse so the user is prevented from restoring the window
from a maximised state. Adding a menubar provides a grab area for
the mouse.

Ideally, the topbar should be a grab area like the menubar is, but
I don't know how to achieve this yet.